requirements.txt
appdirs==1.4.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
conda==4.3.16
distlib==0.3.1
filelock==3.0.12
idna==2.10
pipenv==2020.8.13
pycosat==0.6.3
PyQt5==5.15.3
PyQt5-Qt==5.15.2
PyQt5-sip==12.8.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
speedtest-cli==2.1.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
